Delete working like an ajax request and the order of the delete is fix

+++ Controllers/favoriteDelete.php
@@ -3,8 +3,10 @@ require "../autoload.php";
 
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
 
-    $id = $_POST['id'] ?? null;
+    $id = $_POST['id'];
 
     $pokemon = Pokemon::find($id);
-    $pokemon->delete();
+    if ($pokemon) {
+        $pokemon->delete();
+    }
 }
